Правка замечапний: суффикс "Async", CancellationToken. + Доработки.

+++ WebApplicationMustToHave/Repository/DbCompositionManager.cs
@@ -4,62 +4,84 @@ using WebApplicationMustToHave.Models;
 
 namespace WebApplicationMustToHave.Repository
 {
+    /// <summary>
+    /// Менеджер произведений (интерфейс)
+    /// </summary>
     public interface IDbCompositionManager
     {
         // Возвращает список всех произведений.
         /// <summary>
         /// Возвращает список всех произведений
         /// </summary>
+        /// <param name="cancellationToken">токен отмены</param>
         /// <returns>список</returns>
-        public Task<List<IViewable>> GetCompositionViewsAsync();
+        public Task<List<IViewable>> GetCompositionViewsAsync(CancellationToken cancellationToken);
 
         /// <summary>
         /// Возвращает список всех произведений отфильтрованный по типу
         /// "Книга" = 1, "Аудиокнига" = 2, "Фильм" = 3, "Песня" = 4
         /// </summary>
         /// <param name="typeId"> Книга = 1, Аудиокнига = 2, Фильм = 3, Песня = 4</param>
+        /// <param name="cancellationToken">токен отмены</param>
         /// <returns>список представлений произведений</returns>
-        public Task<List<IViewable>> GetCompositionViewsByTypeAsync(int typeId);
+        public Task<List<IViewable>> GetCompositionViewsByTypeAsync(int typeId, CancellationToken cancellationToken);
 
         /// <summary>
         /// Возвращает список всех произведений
         /// </summary>
+        /// <param name="cancellationToken">токен отмены</param>
         /// <returns>список всех произведений</returns>
-        public Task<List<IDbComposition>> GetCompositionsAsync();
+        public Task<List<IDbComposition>> GetCompositionsAsync(CancellationToken cancellationToken);
 
         /// <summary>
         /// Возвращает список всех произведений отфильтрованный по типу
         /// </summary>
         /// <param name="typeId"> Книга = 1, Аудиокнига = 2, Фильм = 3, Песня = 4</param>
+        /// <param name="cancellationToken">токен отмены</param>
         /// <returns>список</returns>
-        public Task<List<IDbComposition>> GetCompositionsByTypeAsync(int typeId);
+        public Task<List<IDbComposition>> GetCompositionsByTypeAsync(int typeId, CancellationToken cancellationToken);
 
         /// <summary>
         /// Возвращает произведение по id.
         /// </summary>
         /// <param name="id">Ключ произведения</param>
+        /// <param name="cancellationToken">токен отмены</param>
         /// <returns>произведение</returns>
-        public Task<IDbComposition?> GetCompositionByIdAsync(int id);
+        public Task<IDbComposition?> GetCompositionByIdAsync(int id, CancellationToken cancellationToken);
 
         /// <summary>
         /// Обновляет произведение.
         /// </summary>
-        public Task UpdateCompositionAsync(IDbComposition dbComposition);
+        /// <param name="dbComposition">произведение</param>
+        /// <param name="cancellationToken">токен отмены</param>
+        public Task UpdateCompositionAsync(IDbComposition dbComposition, CancellationToken cancellationToken);
 
         /// <summary>
         /// Удаляет произведение.
         /// </summary>
-        public Task DeleteCompositionByIdAsync(int id);
+        /// <param name="id">Ключ произведения</param>
+        /// <param name="cancellationToken">токен отмены</param>
+        public Task DeleteCompositionByIdAsync(int id, CancellationToken cancellationToken);
 
         /// <summary>
         /// Добавляет произведение.
         /// </summary>
-        public Task AddCompositionAsync(IDbComposition dbComposition);
+        /// <param name="dbComposition">произведение</param>
+        /// <param name="cancellationToken">токен отмены</param>
+        public Task AddCompositionAsync(IDbComposition dbComposition, CancellationToken cancellationToken);
 
-        public Task<DbCompositionType?> GetCompositionTypeById(uint typeId);
+        /// <summary>
+        /// Получить тип произведения
+        /// </summary>
+        /// <param name="typeId">Id произведения</param>
+        /// <param name="cancellationToken">токен отмены</param>
+        /// <returns>тип произведения</returns>
+        public Task<DbCompositionType?> GetCompositionTypeByIdAsync(uint typeId, CancellationToken cancellationToken);
     }
 
-
+    /// <summary>
+    /// Менеджер произведений
+    /// </summary>
     public class DbCompositionManager: IDbCompositionManager
     {
         private readonly AppDbContext _db;
@@ -74,9 +96,11 @@ namespace WebApplicationMustToHave.Repository
         /// <summary>
         /// Возвращает список всех произведений
         /// </summary>
+        /// <param name="cancellationToken">токен отмены</param>
         /// <returns>список</returns>
-        public async Task<List<IViewable>> GetCompositionViewsAsync()
+        public async Task<List<IViewable>> GetCompositionViewsAsync(CancellationToken cancellationToken)
         {
+            if (cancellationToken.IsCancellationRequested) return [];
             if (_db.CompositionTypes == null | _db.CompositionTypes.Local.Count == 0) _db.CompositionTypes?.Load();
             List<IViewable> compositions = await _db.Compositions.Where(c => c != null).Select(c => (c as IViewable)!).ToListAsync();
             return compositions ?? [];
@@ -88,9 +112,11 @@ namespace WebApplicationMustToHave.Repository
         /// Возвращает список всех произведений отфильтрованный по типу
         /// </summary>
         /// <param name="typeId"> Книга = 1, Аудиокнига = 2, Фильм = 3, Песня = 4</param>
+        /// <param name="cancellationToken">токен отмены</param>
         /// <returns>список представлений произведений</returns>
-        public async Task<List<IViewable>> GetCompositionViewsByTypeAsync(int typeId)
+        public async Task<List<IViewable>> GetCompositionViewsByTypeAsync(int typeId, CancellationToken cancellationToken)
         {
+            if (cancellationToken.IsCancellationRequested) return [];
             return await _db.Compositions.Where(c => c != null && c.Type != null && c.Type.Id == typeId).Select(c => (c as IViewable)!).ToListAsync() ?? [];
         }
 
@@ -99,9 +125,11 @@ namespace WebApplicationMustToHave.Repository
         /// <summary>
         /// Возвращает список всех произведений
         /// </summary>
+        /// <param name="cancellationToken">токен отмены</param>
         /// <returns>список всех произведений</returns>
-        public async Task<List<IDbComposition>> GetCompositionsAsync()
+        public async Task<List<IDbComposition>> GetCompositionsAsync(CancellationToken cancellationToken)
         {
+            if (cancellationToken.IsCancellationRequested) return [];
             return await _db.Compositions.Select(c => c as IDbComposition).ToListAsync() ?? [];
         }
 
@@ -111,9 +139,11 @@ namespace WebApplicationMustToHave.Repository
         /// Возвращает список всех произведений отфильтрованный по типу
         /// </summary>
         /// <param name="typeId"> Книга = 1, Аудиокнига = 2, Фильм = 3, Песня = 4</param>
+        /// <param name="cancellationToken">токен отмены</param>
         /// <returns>список</returns>
-        public async Task<List<IDbComposition>> GetCompositionsByTypeAsync(int typeId)
+        public async Task<List<IDbComposition>> GetCompositionsByTypeAsync(int typeId, CancellationToken cancellationToken)
         {
+            if (cancellationToken.IsCancellationRequested) return [];
             return await _db.Compositions.Where(c => c.Type != null && c.Type.Id == typeId).Select(c => c as IDbComposition).ToListAsync() ?? [];
         }
 
@@ -121,24 +151,29 @@ namespace WebApplicationMustToHave.Repository
         /// Возвращает произведение по id.
         /// </summary>
         /// <param name="id">Ключ произведения</param>
+        /// <param name="cancellationToken">токен отмены</param>
         /// <returns>произведение</returns>
-        public async Task<IDbComposition?> GetCompositionByIdAsync(int id)
+        public async Task<IDbComposition?> GetCompositionByIdAsync(int id, CancellationToken cancellationToken)
         {
+            if (cancellationToken.IsCancellationRequested) return null;
             IDbComposition dbComposition = await _db.Compositions.Where(c => c.Type != null && c.Id == id).Select(c => c as IDbComposition).FirstOrDefaultAsync();
             if (dbComposition == null) return null;
-            if (dbComposition.Type == null) dbComposition.Type = await GetCompositionTypeById(dbComposition.DbCompositionTypeId);
+            if (dbComposition.Type == null) dbComposition.Type = await GetCompositionTypeByIdAsync(dbComposition.DbCompositionTypeId, cancellationToken);
             return dbComposition;
         }
 
         /// <summary>
         /// Обновляет произведение.
         /// </summary>
-        public async Task UpdateCompositionAsync(IDbComposition dbComposition)
+        /// <param name="dbComposition">произведение</param>
+        /// <param name="cancellationToken">токен отмены</param>
+        public async Task UpdateCompositionAsync(IDbComposition dbComposition, CancellationToken cancellationToken)
         {
             if (dbComposition != null)
             {
+                if (cancellationToken.IsCancellationRequested) return;
                 _db.Compositions.Update((dbComposition as DbComposition)!);
-                await _db.SaveChangesAsync(new CancellationTokenSource().Token);
+                await _db.SaveChangesAsync(cancellationToken);
             }
             else Console.WriteLine("!!!Warning UpdateCompositionAsync() dbComposition = null");
         }
@@ -146,13 +181,16 @@ namespace WebApplicationMustToHave.Repository
         /// <summary>
         /// Удаляет произведение.
         /// </summary>
-        public async Task DeleteCompositionByIdAsync(int id)
+        /// <param name="id">Ключ произведения</param>
+        /// <param name="cancellationToken">токен отмены</param>
+        public async Task DeleteCompositionByIdAsync(int id, CancellationToken cancellationToken)
         {
             DbComposition? dbComposition = _db.Compositions.FirstOrDefault(c => c.Id == id);
             if (dbComposition != null)
             {
+                if (cancellationToken.IsCancellationRequested) return;
                 _db.Compositions.Remove(dbComposition);
-                await _db.SaveChangesAsync(new CancellationTokenSource().Token);
+                await _db.SaveChangesAsync(cancellationToken);
             }
             else Console.WriteLine("!!!Warning DeleteCompositionByIdAsync() dbComposition = null");
         }
@@ -160,22 +198,32 @@ namespace WebApplicationMustToHave.Repository
         /// <summary>
         /// Добавляет произведение.
         /// </summary>
-        public async Task AddCompositionAsync(IDbComposition dbComposition)
+        /// <param name="dbComposition">произведение</param>
+        /// <param name="cancellationToken">токен отмены</param>
+        public async Task AddCompositionAsync(IDbComposition dbComposition, CancellationToken cancellationToken)
         {
             if (dbComposition != null)
             {
+                if (cancellationToken.IsCancellationRequested) return;
                 if (dbComposition.Id == 0)
                 {
                     dbComposition.Id = _db.Compositions.Select(c => c.Id).Max()+1;
                 }
                 _db.Compositions.Add((dbComposition as DbComposition)!);
-                await _db.SaveChangesAsync(new CancellationTokenSource().Token);
+                await _db.SaveChangesAsync(cancellationToken);
             }
             else Console.WriteLine("!!!Warning AddCompositionAsync() dbComposition = null");
         }
 
-        public async Task<DbCompositionType?> GetCompositionTypeById(uint typeId)
+        /// <summary>
+        /// Получить тип произведения
+        /// </summary>
+        /// <param name="typeId">Id произведения</param>
+        /// <param name="cancellationToken">токен отмены</param>
+        /// <returns>тип произведения</returns>
+        public async Task<DbCompositionType?> GetCompositionTypeByIdAsync(uint typeId, CancellationToken cancellationToken)
         {
+            if (cancellationToken.IsCancellationRequested) return null;
             return await _db.CompositionTypes.FirstOrDefaultAsync(c => c.Id == typeId);
         }
